Add existing file

+++ app/main/service/employee_service.py
@@ -5,7 +5,7 @@ from app.main.model.employee import Employee
 
 def create_emp():
     emp_name = request.json['emp_name']
-    email = request.json['emaidell']
+    email = request.json['email']
     new_emp = Employee(emp_name, email)
 
     db.session.add(new_emp)
@@ -19,22 +19,31 @@ def get_emps():
 
 
 def get_emp(id):
-    return Employee.query.get(id)
+    emp = Employee.query.get(id)
+    return emp if emp else {'emp_name': 'No emp with given id'}
 
 
 def update_emp(id):
     emp = Employee.query.get(id)
-    emp_name = request.json['emp_name']
-    email = request.json['email']
-    emp.emp_name = emp_name
-    emp.email = email
-    db.session.commit()
-    return emp
+    if emp:
+        emp_name = request.json['emp_name']
+        email = request.json['email']
+        emp.emp_name = emp_name
+        emp.email = email
+        db.session.commit()
+        return emp
+
+    return {'emp_name': 'No emp with given id'}
 
 
 def delete_emp(id):
     emp = Employee.query.get(id)
-    db.session.delete(emp)
-    db.session.commit()
-    return emp
+    if emp:
+        db.session.delete(emp)
+        db.session.commit()
+        return emp
+    return {'emp_name': 'No emp with given id'}
+
+
+
